pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "A RAG search engine powered by SQLite and sqlite-vec"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "jinja2>=3.1.6",
    "mistune>=3.2.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pysqlite3-binary>=0.5.4.post2",
    "pyyaml>=6.0.3",
    "sqlite-vec>=0.1.6",
    "transformers>=5.2.0",
    "txtai>=9.5.0",
    "uvicorn>=0.41.0",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
]

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "--ignore=tests/manual"
